requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt src/requirements/prod.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
dj-database-url==2.3.0
    # via -r src/requirements/prod.in
django==5.1.6
    # via
    #   -r src/requirements/prod.in
    #   dj-database-url
    #   django-cors-headers
django-cors-headers==4.7.0
    # via -r src/requirements/prod.in
psycopg[binary]==3.2.4
    # via -r src/requirements/prod.in
psycopg-binary==3.2.4
    # via psycopg
python-decouple==3.8
    # via -r src/requirements/prod.in
sqlparse==0.5.3
    # via django
typing-extensions==4.12.2
    # via
    #   dj-database-url
    #   psycopg
whitenoise==6.9.0
    # via -r src/requirements/prod.in
